requirements.txt
ansible==6.6.0
molecule-vagrant==1.0.0    
pytest==7.2.0    
pytest-testinfra==6.8.0